+++ app/src/main/java/com/teamnexters/lastwednesday/fragment/HistoryFragment.java
@@ -1,4 +1,4 @@
-package com.teamnexters.lastwednesday.Fragment;
+package com.teamnexters.lastwednesday.fragment;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -11,15 +11,13 @@ import android.view.ViewGroup;
  * Created by JY on 2018-01-10.
  */
 
-public class MainFragment extends Fragment {
+public class HistoryFragment extends Fragment {
 
-
-    public static MainFragment newInstance() {
-        MainFragment fragment = new MainFragment();
+    public static HistoryFragment newInstance() {
+        HistoryFragment fragment = new HistoryFragment();
 
         return fragment;
     }
-
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
