+++ app/src/main/java/com/teamnexters/lastwednesday/fragment/MainFragment.java
@@ -1,4 +1,4 @@
-package com.teamnexters.lastwednesday.Fragment;
+package com.teamnexters.lastwednesday.fragment;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -11,14 +11,15 @@ import android.view.ViewGroup;
  * Created by JY on 2018-01-10.
  */
 
-public class TicketFragment extends Fragment {
+public class MainFragment extends Fragment {
 
 
-    public static TicketFragment newInstance() {
-        TicketFragment fragment = new TicketFragment();
+    public static MainFragment newInstance() {
+        MainFragment fragment = new MainFragment();
 
         return fragment;
     }
+
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
